Aplicando ORM com Sequelize.

+++ services/product.service.js
@@ -1,8 +1,9 @@
 import ProductRepository from "../repositories/product.repository.js";
 import SupplierRepository from "../repositories/supplier.repository.js";
+import SaleRepository from "../repositories/sale.repository.js";
 
 async function createProduct(product) {
-  if (await SupplierRepository.getSupplier(product.supplier_id)) {
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
     return await ProductRepository.inserProduct(product);
   }
   throw new Error("O supplier não existe.");
@@ -17,11 +18,15 @@ async function getProduct(id) {
 }
 
 async function deleteProduct(id) {
+  const sales = await SaleRepository.getSaleByProductId(id);
+  if (sales.length > 0) {
+    throw new Error("Não é possível excluir o produto pois ele já foi vendido");
+  }
   await ProductRepository.deleteProduct(id);
 }
 
 async function updateProduct(product) {
-  if (await SupplierRepository.getSupplier(product.supplier_id)) {
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
     return await ProductRepository.updateProduct(product);
   }
   throw new Error("O supplier não existe.");
